Linter is setup
