+++ JS/index.js
@@ -1,8 +1,5 @@
-import data1 from '../JS/data.js';
-import data3 from '../JS/data2.js';
-
-console.log(data1);
-console.log(data3);
+import data1 from '../data.js';
+import data3 from '../data2.js';
 
 const body = document.getElementById('body');
 
@@ -35,45 +32,45 @@ headerTopUl.style.gap = '1rem';
 headerTopUl.style.marginRight = '10%';
 headerTopUl.setAttribute('class', 'header-top-ul');
 for (let i = 0; i < 5; i += 1) {
-    if (i === 0) {
-        const headerTopUlLi = document.createElement('img');
-        headerTopUlLi.style.display = 'flex';
-        headerTopUlLi.setAttribute('class', 'header-top-ul-li');
-        headerTopUlLi.src = './Images/facebook-icon.svg';
-        headerTopUlLi.style.height = '20px';
-        headerTopUl.appendChild(headerTopUlLi);
-    } else if (i === 1) {
-        const headerTopUlLi = document.createElement('img');
-        headerTopUlLi.style.display = 'flex';
-        headerTopUlLi.setAttribute('class', 'header-top-ul-li');
-        headerTopUlLi.src = './Images/twitter.svg';
-        headerTopUlLi.style.height = '20px';
-        headerTopUl.appendChild(headerTopUlLi);
-    } else if (i === 2) {
-        const headerTopUlLi = document.createElement('li');
-        headerTopUlLi.style.display = 'flex';
-        headerTopUlLi.setAttribute('class', 'header-top-ul-li');
-        headerTopUlLi.textContent = 'English';
-        headerTopUlLi.style.color = '#fff';
-        headerTopUlLi.style.fontFamily = 'Lato, sans-serif';
-        headerTopUl.appendChild(headerTopUlLi);
-    } else if (i === 3) {
-        const headerTopUlLi = document.createElement('li');
-        headerTopUlLi.style.display = 'flex';
-        headerTopUlLi.setAttribute('class', 'header-top-ul-li');
-        headerTopUlLi.textContent = 'My Page';
-        headerTopUlLi.style.color = '#fff';
-        headerTopUlLi.style.fontFamily = 'Lato, sans-serif';
-        headerTopUl.appendChild(headerTopUlLi);
-    } else if (i === 4) {
-        const headerTopUlLi = document.createElement('li');
-        headerTopUlLi.style.display = 'flex';
-        headerTopUlLi.setAttribute('class', 'header-top-ul-li');
-        headerTopUlLi.textContent = 'Logout';
-        headerTopUlLi.style.color = '#fff';
-        headerTopUlLi.style.fontFamily = 'Lato, sans-serif';
-        headerTopUl.appendChild(headerTopUlLi);
-    }
+  if (i === 0) {
+    const headerTopUlLi = document.createElement('img');
+    headerTopUlLi.style.display = 'flex';
+    headerTopUlLi.setAttribute('class', 'header-top-ul-li');
+    headerTopUlLi.src = './Images/facebook-icon.svg';
+    headerTopUlLi.style.height = '20px';
+    headerTopUl.appendChild(headerTopUlLi);
+  } else if (i === 1) {
+    const headerTopUlLi = document.createElement('img');
+    headerTopUlLi.style.display = 'flex';
+    headerTopUlLi.setAttribute('class', 'header-top-ul-li');
+    headerTopUlLi.src = './Images/twitter.svg';
+    headerTopUlLi.style.height = '20px';
+    headerTopUl.appendChild(headerTopUlLi);
+  } else if (i === 2) {
+    const headerTopUlLi = document.createElement('li');
+    headerTopUlLi.style.display = 'flex';
+    headerTopUlLi.setAttribute('class', 'header-top-ul-li');
+    headerTopUlLi.textContent = 'English';
+    headerTopUlLi.style.color = '#fff';
+    headerTopUlLi.style.fontFamily = 'Lato, sans-serif';
+    headerTopUl.appendChild(headerTopUlLi);
+  } else if (i === 3) {
+    const headerTopUlLi = document.createElement('li');
+    headerTopUlLi.style.display = 'flex';
+    headerTopUlLi.setAttribute('class', 'header-top-ul-li');
+    headerTopUlLi.textContent = 'My Page';
+    headerTopUlLi.style.color = '#fff';
+    headerTopUlLi.style.fontFamily = 'Lato, sans-serif';
+    headerTopUl.appendChild(headerTopUlLi);
+  } else if (i === 4) {
+    const headerTopUlLi = document.createElement('li');
+    headerTopUlLi.style.display = 'flex';
+    headerTopUlLi.setAttribute('class', 'header-top-ul-li');
+    headerTopUlLi.textContent = 'Logout';
+    headerTopUlLi.style.color = '#fff';
+    headerTopUlLi.style.fontFamily = 'Lato, sans-serif';
+    headerTopUl.appendChild(headerTopUlLi);
+  }
 }
 headerTop.appendChild(headerTopUl);
 
@@ -87,11 +84,11 @@ headerBottom.style.marginBottom = '1rem';
 headerBottom.style.gap = '1rem';
 headerBottom.setAttribute('class', 'header-bottom');
 
-const headerBottomDiv= document.createElement('div');
+const headerBottomDiv = document.createElement('div');
 headerBottomDiv.style.display = 'flex';
-headerBottomDiv.style.width = '100%'
+headerBottomDiv.style.width = '100%';
 headerBottomDiv.style.marginLeft = '10%';
-headerBottomDiv.style.gap = '1rem'
+headerBottomDiv.style.gap = '1rem';
 headerBottomDiv.setAttribute('class', 'header-bottom-div');
 
 const headerBottomDiv1 = document.createElement('div');
@@ -135,41 +132,31 @@ headerBottomUl.style.gap = '2rem';
 headerBottomUl.style.justifyContent = 'end';
 headerBottomUl.setAttribute('class', 'header-bottom-ul');
 for (let i = 0; i < 6; i += 1) {
-    const headerBottomUlLi = document.createElement('li');
-    headerBottomUlLi.style.display = 'flex';
-    headerBottomUlLi.setAttribute('class', 'header-top-ul-li');
-    if (i === 0) {
+  const headerBottomUlLi = document.createElement('li');
+  headerBottomUlLi.style.display = 'flex';
+  headerBottomUlLi.setAttribute('class', 'header-top-ul-li');
+  if (i === 0) {
+    headerBottomUlLi.textContent = 'About';
+    headerBottomUlLi.style.color = '#ec5242';
 
-        headerBottomUlLi.textContent = 'About';
-        headerBottomUlLi.style.color = '#ec5242';
-
-        headerBottomUlLi.addEventListener('click', () => {
-            location.href = '/about.html';
-          });
-        
-    } else if (i === 1) {
-        headerBottomUlLi.textContent = 'Program';
-
-    } else if (i === 2) {
-
-        headerBottomUlLi.textContent = 'Join';
-
-    } else if (i === 3) {
-
-        headerBottomUlLi.textContent = 'Sponsor';
-
-    } else if (i === 4) {
-
-        headerBottomUlLi.textContent = 'News';
-
-    } else if (i === 5) {
-
-        headerBottomUlLi.textContent = 'CC Campain';
-        headerBottomUlLi.style.padding = '0.5rem';
-        headerBottomUlLi.style.border = '3px solid #ec5242';
-        headerBottomUlLi.style.color = '#ec5242';
-    }
-    headerBottomUl.appendChild(headerBottomUlLi);
+    headerBottomUlLi.addEventListener('click', () => {
+      window.location.href = '/about.html';
+    });
+  } else if (i === 1) {
+    headerBottomUlLi.textContent = 'Program';
+  } else if (i === 2) {
+    headerBottomUlLi.textContent = 'Join';
+  } else if (i === 3) {
+    headerBottomUlLi.textContent = 'Sponsor';
+  } else if (i === 4) {
+    headerBottomUlLi.textContent = 'News';
+  } else if (i === 5) {
+    headerBottomUlLi.textContent = 'CC Campain';
+    headerBottomUlLi.style.padding = '0.5rem';
+    headerBottomUlLi.style.border = '3px solid #ec5242';
+    headerBottomUlLi.style.color = '#ec5242';
+  }
+  headerBottomUl.appendChild(headerBottomUlLi);
 }
 headerBottom.appendChild(headerBottomUl);
 
@@ -193,7 +180,7 @@ body1txt1.style.display = 'flex';
 body1txt1.setAttribute('class', 'body1-txt1');
 body1txt1.textContent = '"Hello! Sharing world"';
 body1txt1.style.color = '#ec5242';
-body1txt1.style.marginTop= '5%';
+body1txt1.style.marginTop = '5%';
 body1txt1.style.fontFamily = 'Lato, sans-serif';
 body1txt1.style.marginLeft = '10%';
 body1txt1.style.fontSize = '2rem';
@@ -296,57 +283,56 @@ body2div1.style.gap = '0.5rem';
 body2div1.setAttribute('class', 'body2-div1');
 
 function fillMainProgram(incomDiv, proName, proImg, proDesc) {
+  const body2Div = document.createElement('div');
+  body2Div.style.display = 'flex';
+  body2Div.style.flexDirection = 'column';
+  body2Div.setAttribute('class', 'body2-div');
+  body2Div.style.backgroundColor = '#272a31';
+  body2Div.style.height = '100%';
+  body2Div.style.opacity = '0.8';
+  body2Div.style.width = '15%';
 
-    const body2Div = document.createElement('div');
-    body2Div.style.display = 'flex';
-    body2Div.style.flexDirection = 'column';
-    body2Div.setAttribute('class', 'body2-div');
-    body2Div.style.backgroundColor = '#272a31';
-    body2Div.style.height = '100%';
-    body2Div.style.opacity = '0.8';
-    body2Div.style.width = '15%';
-    
-    const body2Divtxt1 = document.createElement('p');
-    body2Divtxt1.style.display = 'flex';
-    body2Divtxt1.style.justifyContent = 'center';
-    body2Divtxt1.setAttribute('class', 'body2-div-txt1');
-    body2Divtxt1.textContent = proName;
-    body2Divtxt1.style.color = '#ec5242';
-    body2Divtxt1.style.fontFamily = 'Lato, sans-serif';
-    body2Divtxt1.style.fontSize = '1rem';
-    body2Divtxt1.style.fontStyle = 'normal';
+  const body2Divtxt1 = document.createElement('p');
+  body2Divtxt1.style.display = 'flex';
+  body2Divtxt1.style.justifyContent = 'center';
+  body2Divtxt1.setAttribute('class', 'body2-div-txt1');
+  body2Divtxt1.textContent = proName;
+  body2Divtxt1.style.color = '#ec5242';
+  body2Divtxt1.style.fontFamily = 'Lato, sans-serif';
+  body2Divtxt1.style.fontSize = '1rem';
+  body2Divtxt1.style.fontStyle = 'normal';
 
-    const body2Divtxt2 = document.createElement('p');
-    body2Divtxt2.style.display = 'flex';
-    body2Divtxt2.style.justifyContent = 'center';
-    body2Divtxt2.setAttribute('class', 'body2-div-txt2');
-    body2Divtxt2.textContent = proDesc;
-    body2Divtxt2.style.color = '#fff';
-    body2Divtxt2.style.margin = '1rem';
-    body2Divtxt2.style.fontFamily = 'Lato, sans-serif';
-    body2Divtxt2.style.fontSize = '0.7rem';
-    body2Divtxt2.style.fontStyle = 'normal';
+  const body2Divtxt2 = document.createElement('p');
+  body2Divtxt2.style.display = 'flex';
+  body2Divtxt2.style.justifyContent = 'center';
+  body2Divtxt2.setAttribute('class', 'body2-div-txt2');
+  body2Divtxt2.textContent = proDesc;
+  body2Divtxt2.style.color = '#fff';
+  body2Divtxt2.style.margin = '1rem';
+  body2Divtxt2.style.fontFamily = 'Lato, sans-serif';
+  body2Divtxt2.style.fontSize = '0.7rem';
+  body2Divtxt2.style.fontStyle = 'normal';
 
-    const body2DivImg = document.createElement('img');
-    body2DivImg.style.display = 'flex';
-    body2DivImg.src = proImg;
-    body2DivImg.style.justifyContent = 'center';
-    body2DivImg.style.height = '2rem';
-    body2DivImg.style.marginTop = '1rem';
-    body2DivImg.style.marginBottom = '1rem';
-    body2DivImg.style.alignItems = 'center';
-    body2DivImg.setAttribute('class', 'setion2-dynamic-img');
+  const body2DivImg = document.createElement('img');
+  body2DivImg.style.display = 'flex';
+  body2DivImg.src = proImg;
+  body2DivImg.style.justifyContent = 'center';
+  body2DivImg.style.height = '2rem';
+  body2DivImg.style.marginTop = '1rem';
+  body2DivImg.style.marginBottom = '1rem';
+  body2DivImg.style.alignItems = 'center';
+  body2DivImg.setAttribute('class', 'setion2-dynamic-img');
 
-    body2Div.appendChild(body2DivImg);
-    body2Div.appendChild(body2Divtxt1);
-    body2Div.appendChild(body2Divtxt2);
-    
-    incomDiv.appendChild(body2Div);
-  }
+  body2Div.appendChild(body2DivImg);
+  body2Div.appendChild(body2Divtxt1);
+  body2Div.appendChild(body2Divtxt2);
+
+  incomDiv.appendChild(body2Div);
+}
 
 for (let i = 0; i < data1.length; i += 1) {
-    fillMainProgram(body2div1, data1[i].programName, data1[i].programImg, data1[i].programDesc);
-  }
+  fillMainProgram(body2div1, data1[i].programName, data1[i].programImg, data1[i].programDesc);
+}
 
 const body2txt2 = document.createElement('p');
 body2txt2.style.display = 'flex';
@@ -413,90 +399,87 @@ sec3Div.style.gap = '0.5rem';
 sec3Div.setAttribute('class', 'section3-div');
 
 function fillFeaturedSpeackers(incomSec3Div, spkrImg, spkrName, spkrTitle, spkrDesc) {
+  const section3Div = document.createElement('div');
+  section3Div.style.display = 'flex';
+  section3Div.setAttribute('class', 'seaction3-each-div');
+  section3Div.style.backgroundColor = '#fff';
+  section3Div.style.height = '100%';
+  section3Div.style.width = '40%';
 
-    const section3Div = document.createElement('div');
-    section3Div.style.display = 'flex';
-    section3Div.setAttribute('class', 'seaction3-each-div');
-    section3Div.style.backgroundColor = '#fff';
-    section3Div.style.height = '100%';
-    section3Div.style.width = '40%';
+  const section3Div1 = document.createElement('div');
+  section3Div1.style.display = 'flex';
+  section3Div1.setAttribute('class', 'seaction3-first-div');
+  section3Div1.style.width = '35%';
+  section3Div1.style.margin = '0.5rem';
 
-    const section3Div1 = document.createElement('div');
-    section3Div1.style.display = 'flex';
-    section3Div1.setAttribute('class', 'seaction3-first-div');
-    section3Div1.style.width = '35%';
-    section3Div1.style.margin = '0.5rem';
+  const section3Div1Img = document.createElement('img');
+  section3Div1Img.style.display = 'flex';
+  section3Div1Img.style.width = '100%';
+  section3Div1Img.src = spkrImg;
+  section3Div1Img.style.justifyContent = 'center';
+  section3Div1Img.style.alignItems = 'center';
+  section3Div1Img.setAttribute('class', 'setion3-first-div-img');
 
-    const section3Div1Img = document.createElement('img');
-    section3Div1Img.style.display = 'flex';
-    section3Div1Img.style.width = '100%';
-    section3Div1Img.src = spkrImg;
-    section3Div1Img.style.justifyContent = 'center';
-    section3Div1Img.style.alignItems = 'center';
-    section3Div1Img.setAttribute('class', 'setion3-first-div-img');
+  const section3Div2 = document.createElement('div');
+  section3Div2.style.display = 'flex';
+  section3Div2.style.flexDirection = 'column';
+  section3Div2.setAttribute('class', 'seaction3-second-div');
+  section3Div2.style.width = '65%';
+  section3Div2.style.margin = '0.5rem';
 
-    const section3Div2 = document.createElement('div');
-    section3Div2.style.display = 'flex';
-    section3Div2.style.flexDirection = 'column';
-    section3Div2.setAttribute('class', 'seaction3-second-div');
-    section3Div2.style.width = '65%';
-    section3Div2.style.margin = '0.5rem';
+  const section3Div2txt1 = document.createElement('p');
+  section3Div2txt1.style.display = 'flex';
+  // section3Div2txt1.style.fontFamily = 'Lato, sans-serif';
+  section3Div2txt1.style.fontFamily = 'COCOGOOSE, sans-serif';
+  section3Div2txt1.setAttribute('class', 'body2-div-txt1');
+  section3Div2txt1.textContent = spkrName;
+  section3Div2txt1.style.fontStyle = 'normal';
+  section3Div2txt1.style.color = '#272a31';
+  section3Div2txt1.style.fontSize = '0.8rem';
 
-    const section3Div2txt1 = document.createElement('p');
-    section3Div2txt1.style.display = 'flex';
-    // section3Div2txt1.style.fontFamily = 'Lato, sans-serif';
-    section3Div2txt1.style.fontFamily= 'COCOGOOSE, sans-serif';
-    section3Div2txt1.setAttribute('class', 'body2-div-txt1');
-    section3Div2txt1.textContent = spkrName;
-    section3Div2txt1.style.fontStyle = 'normal';
-    section3Div2txt1.style.color = '#272a31';
-    section3Div2txt1.style.fontSize = '0.8rem';
+  const section3Div2txt2 = document.createElement('p');
+  section3Div2txt2.style.display = 'flex';
+  section3Div2txt2.style.fontFamily = 'Lato, sans-serif';
+  // section3Div2txt2.style.fontFamily= 'COCOGOOSE, sans-serif';
+  section3Div2txt2.setAttribute('class', 'body2-div-txt2');
+  section3Div2txt2.textContent = spkrTitle;
+  section3Div2txt2.style.fontStyle = 'italic';
+  section3Div2txt2.style.color = '#ec5242';
+  section3Div2txt2.style.fontSize = '0.7rem';
 
-    const section3Div2txt2 = document.createElement('p');
-    section3Div2txt2.style.display = 'flex';
-    section3Div2txt2.style.fontFamily = 'Lato, sans-serif';
-    // section3Div2txt2.style.fontFamily= 'COCOGOOSE, sans-serif';
-    section3Div2txt2.setAttribute('class', 'body2-div-txt2');
-    section3Div2txt2.textContent = spkrTitle;
-    section3Div2txt2.style.fontStyle = 'italic';
-    section3Div2txt2.style.color = '#ec5242';
-    section3Div2txt2.style.fontSize = '0.7rem';
+  const section3Hr = document.createElement('hr');
+  section3Hr.style.display = 'flex';
+  section3Hr.style.border = '1px solid #d3d3d3';
+  section3Hr.style.width = '8%';
+  section3Hr.style.marginBottom = '0.5rem';
+  section3Hr.style.marginTop = '0.5rem';
 
-    const section3Hr = document.createElement('hr');
-    section3Hr.style.display = 'flex';
-    section3Hr.style.border = '1px solid #d3d3d3';
-    section3Hr.style.width = '8%';
-    section3Hr.style.marginBottom = '0.5rem';
-    section3Hr.style.marginTop = '0.5rem';
+  const section3Div2txt3 = document.createElement('p');
+  section3Div2txt3.style.display = 'flex';
+  section3Div2txt3.style.justifyContent = 'center';
+  section3Div2txt3.style.fontFamily = 'Lato, sans-serif';
+  section3Div2txt3.setAttribute('class', 'body2-div-txt3');
+  section3Div2txt3.textContent = spkrDesc;
+  section3Div2txt3.style.fontStyle = 'normal';
+  section3Div2txt3.style.color = '#272a31';
+  section3Div2txt3.style.fontSize = '0.8rem';
 
-    const section3Div2txt3 = document.createElement('p');
-    section3Div2txt3.style.display = 'flex';
-    section3Div2txt3.style.justifyContent = 'center';
-    section3Div2txt3.style.fontFamily = 'Lato, sans-serif';
-    section3Div2txt3.setAttribute('class', 'body2-div-txt3');
-    section3Div2txt3.textContent = spkrDesc;
-    section3Div2txt3.style.fontStyle = 'normal';
-    section3Div2txt3.style.color = '#272a31';
-    section3Div2txt3.style.fontSize = '0.8rem';
+  section3Div2.appendChild(section3Div2txt1);
+  section3Div2.appendChild(section3Div2txt2);
+  section3Div2.appendChild(section3Hr);
+  section3Div2.appendChild(section3Div2txt3);
 
-
-    section3Div2.appendChild(section3Div2txt1);
-    section3Div2.appendChild(section3Div2txt2);
-    section3Div2.appendChild(section3Hr);
-    section3Div2.appendChild(section3Div2txt3);
-
-    
-    // section3Div1.appendChild(section3Div1Img1);
-    section3Div1.appendChild(section3Div1Img);
-    section3Div.appendChild(section3Div1);
-    section3Div.appendChild(section3Div2);
-    incomSec3Div.appendChild(section3Div);
-
+  // section3Div1.appendChild(section3Div1Img1);
+  section3Div1.appendChild(section3Div1Img);
+  section3Div.appendChild(section3Div1);
+  section3Div.appendChild(section3Div2);
+  incomSec3Div.appendChild(section3Div);
 }
 
 for (let i = 0; i < data3.length; i += 1) {
-    fillFeaturedSpeackers(sec3Div, data3[i].speackerImg, data3[i].speackerName, data3[i].speackerTitle, data3[i].speackerDesc);
-  }
+  fillFeaturedSpeackers(sec3Div, data3[i].speackerImg, data3[i].speackerName,
+    data3[i].speackerTitle, data3[i].speackerDesc);
+}
 
 sec3TxtDiv.appendChild(body3txt1);
 sec3TxtDiv.appendChild(body3Hr);
@@ -510,7 +493,6 @@ footerSection.style.flexDirection = 'column';
 footerSection.setAttribute('class', 'section-footer');
 footerSection.style.height = '20%';
 footerSection.style.backgroundColor = '#f2f2f2';
-
 
 const footerSectionPart1 = document.createElement('div');
 footerSectionPart1.style.display = 'flex';
@@ -544,34 +526,32 @@ footerHr.style.display = 'flex';
 footerHr.style.justifyContent = 'center';
 footerHr.style.width = '3%';
 
-
 const footer1Div2 = document.createElement('div');
 footer1Div2.style.display = 'flex';
 footer1Div2.style.width = '100%';
 footer1Div2.style.justifyContent = 'center';
 footer1Div2.style.alignItems = 'center';
 footer1Div2.style.marginLeft = '3rem';
-footer1Div2.style.marginRight= '3rem';
-footer1Div2.style.marginBottom= '3rem';
-footer1Div2.style.marginTop= '3rem';
+footer1Div2.style.marginRight = '3rem';
+footer1Div2.style.marginBottom = '3rem';
+footer1Div2.style.marginTop = '3rem';
 footer1Div2.style.gap = '5rem';
 footer1Div2.setAttribute('class', 'footer-div1');
 
 for (let i = 0; i < 5; i += 1) {
+  const footer1Div = document.createElement('div');
+  footer1Div.style.display = 'flex';
+  footer1Div.style.justifyContent = 'center';
+  footer1Div.style.alignItems = 'center';
+  footer1Div.setAttribute('class', 'footer1-div');
 
-const footer1Div = document.createElement('div');
-footer1Div.style.display = 'flex';
-footer1Div.style.justifyContent = 'center';
-footer1Div.style.alignItems = 'center';
-footer1Div.setAttribute('class', 'footer1-div');
+  const footer1Txt = document.createElement('p');
+  footer1Txt.style.display = 'flex';
+  footer1Txt.style.justifyContent = 'center';
+  footer1Txt.style.alignItems = 'center';
+  footer1Txt.setAttribute('class', 'footer1-txt');
 
-const footer1Txt = document.createElement('p');
-footer1Txt.style.display = 'flex';
-footer1Txt.style.justifyContent = 'center';
-footer1Txt.style.alignItems = 'center';
-footer1Txt.setAttribute('class', 'footer1-txt');
-
-if (i === 0) {
+  if (i === 0) {
     footer1Txt.textContent = 'mozilla';
     footer1Txt.style.fontFamily = 'Lato, sans-serif';
     footer1Txt.style.fontSize = '2rem';
@@ -579,26 +559,25 @@ if (i === 0) {
     footer1Txt.style.fontStyle = 'bold';
     footer1Txt.style.color = '#ccc';
     // body1txt2.style.fontFamily = 'COCOGOOSE, sans-serif';
-} else if (i === 1) {
+  } else if (i === 1) {
     footer1Txt.textContent = 'Google';
     footer1Txt.style.fontFamily = 'sans-serif';
     footer1Txt.style.fontSize = '1.5rem';
     footer1Txt.style.fontStyle = 'normal';
     footer1Txt.style.color = '#d3d3d3';
-} else if (i === 2) {
+  } else if (i === 2) {
     footer1Txt.textContent = 'NAVER';
     footer1Txt.style.fontSize = '2rem';
     footer1Txt.style.fontWeight = '700';
     footer1Txt.style.fontStyle = 'bold';
     footer1Txt.style.color = '#ccc';
     footer1Txt.style.fontFamily = 'COCOGOOSE, sans-serif';
-} else if (i === 3) {
+  } else if (i === 3) {
     footer1Txt.style.fontSize = '2rem';
     footer1Txt.style.fontStyle = 'normal';
     footer1Txt.style.color = '#ccc';
     footer1Txt.textContent = 'daumkakao';
-} else if (i === 4) {
-
+  } else if (i === 4) {
     footer1Txt.style.fontSize = '2rem';
     footer1Txt.style.fontStyle = 'normal';
     footer1Txt.style.color = '#ccc';
@@ -614,13 +593,12 @@ if (i === 0) {
     footer1Div1Img.src = './Images/airbnb-icon.svg';
 
     footer1Div.appendChild(footer1Div1Img);
+  }
+
+  footer1Div.appendChild(footer1Txt);
+
+  footer1Div2.appendChild(footer1Div);
 }
-
-footer1Div.appendChild(footer1Txt);
-
-footer1Div2.appendChild(footer1Div);
-}
-
 
 footer1Div1.appendChild(footerTxt0);
 footer1Div1.appendChild(footerHr);
@@ -640,8 +618,8 @@ footerSectionPart2.style.marginBottom = '2rem';
 footerSectionPart2.style.justifyContent = 'center';
 footerSectionPart2.style.alignItems = 'center';
 footerSectionPart2.style.gap = '2rem';
-footerSectionPart2.style.marginBottom= '3rem';
-footerSectionPart2.style.marginTop= '3rem';
+footerSectionPart2.style.marginBottom = '3rem';
+footerSectionPart2.style.marginTop = '3rem';
 footerSectionPart2.setAttribute('class', 'footer-section-part2');
 
 const footer2Div1 = document.createElement('div');
