+++ data.js
@@ -1,0 +1,49 @@
+const data1 = [
+  {
+    id: 1,
+    programName: 'Youtube',
+    programDesc: 'Lorem Ipsum is simply dd not only five cc typesetting, remaining essentonly five cc typesetting, remaining essent',
+    programImg: './Images/Vector1.svg',
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 2,
+    programName: 'Facebook',
+    programDesc: 'Lorem Ipsum is simply dd not only five cc typesetting, remaining essentonly five cc typesetting, remaining essent',
+    programImg: './Images/Vector2.svg',
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 3,
+    programName: 'Twitter',
+    programDesc: 'Lorem Ipsum is simply dd not only five cc typesetting, remaining essentonly five cc typesetting, remaining essent',
+    programImg: './Images/Vector3.svg',
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 4,
+    programName: 'Telegram',
+    programDesc: 'Lorem Ipsum is simply dd not only five cc typesetting, remaining essentonly five cc typesetting, remaining essent',
+    programImg: './Images/Vector4.svg',
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 5,
+    programName: 'Instagram',
+    programDesc: 'Lorem Ipsum is simply dd not only five cc typesetting, remaining essentonly five cc typesetting, remaining essent',
+    programImg: './Images/Vector5.svg',
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+];
+
+export default data1;
