+++ data2.js
@@ -1,0 +1,58 @@
+const data3 = [
+  {
+    id: 1,
+    speackerName: 'Yochai Benkler',
+    speackerDesc: 'Yochai Benkler Lorem Ipsum is simply dd not only five cc typesetting, remaining essentonly five cc typesetting, remaining essent',
+    speackerImg: '../Images/pic3.jpg',
+    speackerTitle: 'Berkman professor of Enterpreneurial Legal Studies at Harvard Low School',
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 2,
+    speackerName: 'Kilnam Chon',
+    speackerDesc: 'Kilnam Chon Lorem Ipsum is simply dummy text of the printing and typesetting andard essent',
+    speackerImg: '../Images/pic2.jpg',
+    speackerTitle: '',
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 3,
+    speackerName: 'Julia Leda',
+    speackerDesc: 'Julia Leda Lorem Ipsum is simply dummy text of the stry. Lhe industry\'s standard dummy text ',
+    speackerImg: '../Images/pic1.jpg',
+    speackerTitle: 'Director of Art Centre Nabi and a board member of CC Korea',
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 4,
+    speackerName: 'SohYeong Noh',
+    speackerDesc: 'SohYeong Noh Lorem Ipsum is simply dummy text of the  Ipsuver since the 1500s',
+    speackerImg: '../Images/pic4.jpg',
+    speackerTitle: 'President of Young pirates of Europe',
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 5,
+    speackerName: 'Lila Tretikov',
+    speackerDesc: 'Lila Tretikov Lorem a galleleap into electronic typesetting, remaining essent',
+    speackerImg: '../Images/pic5.jpg',
+    speackerTitle: 'Executive Director of the Wikimedia Foundation',
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+  {
+    id: 6,
+    speackerName: 'Ryan Merkley',
+    speackerDesc: 'Ryan Merkley Lorem a galley of type and scap into electronic typesetting, remaining essent',
+    speackerImg: '../Images/pic6.jpg',
+    speackerTitle: 'CEO of Creative Commons, ex COO of the Mozila Foundation',
+    liveVersion: 'https://abiy006.github.io/',
+    source: 'https://github.com/abiy006/Abiy-Portfolio',
+  },
+];
+
+export default data3;
